+++ aofl-js-packages/web-components/aofl-multiselect-list/template.js
@@ -1,0 +1,8 @@
+export const template = (context, html) => html`
+<style>
+:host {
+  display: block;
+}
+</style>
+<slot></slot>
+`;